requirements.txt
torch>=2.0
torchvision
matplotlib
numpy
Pillow
